requirements.txt
# Core dependencies
pandas==2.1.3
numpy==1.24.3
boto3==1.34.129
requests==2.31.0
pyspark==3.5.0
great-expectations==0.18.5
kubernetes==28.1.0
python-dotenv==1.0.0
apache-airflow-providers-cncf-kubernetes==8.4.0
apache-airflow-providers-http==4.9.0

# Data processing
pyarrow==14.0.1
fastparquet==2023.10.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redshift-connector==2.0.915

# AWS services
awswrangler==3.8.0
s3fs==2024.6.1

# Data quality
jsonschema==4.19.2
cerberus==1.3.5

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Configuration
pydantic==2.5.0
PyYAML==6.0.1

# Utilities
click==8.1.7
tqdm==4.66.1
schedule==1.2.0

# Visualization
tableau-api-lib==0.1.46
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0